requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
numpy==2.4.3
pandas==3.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
starlette==0.52.1
threadpoolctl==3.6.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
